Change precision for DecimalEx operations to 20 due error in the results

+++ math_engine/src/main.rs
@@ -1,6 +1,6 @@
 use math_engine::evaluator::Evaluator;
 
 fn main(){
-    let evaluator : Evaluator<i32> = Evaluator::new();
+    let evaluator : Evaluator<f64> = Evaluator::new();
     println!("{:?}", evaluator.eval("Random(0, 100)"));
 }
